requirements.txt
requests
requests_oauthlib
python-dotenv
jupyter
PyYAML
